requirements.txt
pycrypto
flask
flask_login
flask_restplus
flask_sqlalchemy